用lightFM设计的推荐系统
LightFM is a Python implementation of a number of popular recommendation
algorithms for both implicit and explicit feedback.

It also makes it possible to incorporate both item and user metadata
into the traditional matrix factorization algorithms. It represents each
user and item as the sum of the latent representations of their
features, thus allowing recommendations to generalise to new items (via
item features) and to new users (via user features).